fixed route so edit container and input parameter endpoints both accessible; importing enum from db server doesn't work
